requirements.txt
python-dotenv
pygame